Imported Upstream version 1.4

+++ src/viewer/ircview.cpp
@@ -291,30 +291,16 @@ struct Burr: public QTextBlockUserData
 
     ~Burr()
     {
-        KDebug::Block myBlock(qPrintable(QString::number((ulong)(this), 16)));
-
-        kDebug() << "~Burr" << (void*)this << _S(m_format) << _S(m_block.blockNumber()) << "deleted";
         m_owner->blockDeleted(this);
         unlink();
     }
 
     void unlink()
     {
-        kDebug() << "====" << (void*)this << m_block;
         if (m_prev)
-        {
-            kDebug() << "prev->next" << (void*)m_next;
             m_prev->m_next = m_next;
-        }
-        else
-            kDebug() << "prev was null";
         if (m_next)
-        {
-            kDebug() << "next";
             m_next->m_prev = m_prev;
-        }
-        else
-            kDebug() << "next was null";
     }
 
     QTextBlock m_block;
@@ -373,9 +359,7 @@ QTextCharFormat IRCView::getFormat(ObjectFormats x)
 
 void IRCView::blockDeleted(Burr* b) //slot
 {
-    DebugBanner;
     Q_ASSERT(b); // this method only to be called from a ~Burr();
-    kDebug() << "blockDeleted" << (void*)(b) << b->m_block << _S(m_lastMarkerLine) << _S(m_rememberLine);
 
     //tracking only the tail
     if (b == m_lastMarkerLine)
@@ -385,66 +369,17 @@ void IRCView::blockDeleted(Burr* b) //slot
         m_rememberLine = 0;
 }
 
-void IRCView::cullMarkedLine(int where, int rem, int add) //slot
+void IRCView::cullMarkedLine(int, int, int) //slot
 {
-    bool showDebug = false;
-    QString output;
-    QDebug d = QDebug(&output);//KDebug(QtDebugMsg, __FILE__, __LINE__, Q_FUNC_INFO)();
-
-    bool merged = (add!=0 && rem !=0); // i have never seen this happen, adds and removes are always separate
-    int blockCount = document()->blockCount();
-    void *view = this;
-    QTextBlock prime = document()->firstBlock();
-
-    d << "================= cullMarkedLine" << _S(view) << _S(where) << _S(rem) << _S(add) << _S(blockCount) << _S(prime.length()) << _S(merged);
-
-    if (prime.length() == 1)
-    {
-        if (document()->blockCount() == 1) //the entire document was wiped. was a signal such a burden? apparently..
-        {
-            showDebug = true;
-            d << "- wipeLineParagraphs()" << (void*)m_rememberLine << (void*)m_lastMarkerLine;
-            wipeLineParagraphs();
-        }
-        else if (document()->characterAt(0).unicode() == 0x2029)
-        {
-            showDebug = true;
-            d << "- only QChar::ParagraphSeparator";
-            // this should never happen, it should be 0xfffc2029
-            if (dynamic_cast<Burr*>(prime.userData()))
-                d << "Burr!" << prime.userData();
-        }
-        else
-        {
-            showDebug = true;
-            QString fc = "0x" + QString::number(document()->characterAt(0).unicode(), 16).rightJustified(4, '0');
-            d << "- block of length 1 but not 2029" << qPrintable(fc);
-        }
-    }
-    else if (prime.length() == 2)
-    {
-        //probably a Burr going to be culled next..
-        showDebug = true;
-        QString fc = "0x" + QString::number(document()->characterAt(0).unicode(), 16).rightJustified(4, '0');
-        QString sc = "0x" + QString::number(document()->characterAt(1).unicode(), 16).rightJustified(4, '0');
-        d << "- prime(2)" << fc << sc;
-    }
-    if (showDebug)
-    {
-        DebugBanner;
-        kDebug() << output;
-    }
+    if (document()->firstBlock().length() == 1 && document()->blockCount() == 1) //the entire document was wiped. was a signal such a burden? apparently..
+        wipeLineParagraphs();
 }
 
 void IRCView::insertMarkerLine() //slot
 {
-    DebugBanner;
-
     //if the last line is already a marker of any kind, skip out
     if (lastBlockIsLine(BlockIsMarker))
         return;
-
-    kDebug() << "ok to insert..";
 
     //the code used to preserve the dirty bit status, but that was never affected by appendLine...
     //maybe i missed something
@@ -453,13 +388,10 @@ void IRCView::insertMarkerLine() //slot
 
 void IRCView::insertRememberLine() //slot
 {
-    DebugBanner;
-
     m_rememberLineDirtyBit = true; // means we're going to append a remember line if some text gets inserted
 
     if (!Preferences::self()->automaticRememberLineOnlyOnTextChange())
     {
-        kDebug();
         appendRememberLine();
     }
 }
@@ -478,8 +410,6 @@ bool IRCView::lastBlockIsLine(int select)
     if (b)
         state = b->m_format;
 
-    kDebug() << _S(state) << _S(select);
-
     if (select == -1)
         return (state == BlockIsRemember || state == BlockIsMarker);
 
@@ -488,8 +418,6 @@ bool IRCView::lastBlockIsLine(int select)
 
 void IRCView::appendRememberLine()
 {
-    DebugBanner;
-
     //clear this now, so that doAppend doesn't double insert
     m_rememberLineDirtyBit = false;
 
@@ -503,15 +431,12 @@ void IRCView::appendRememberLine()
         voidLineBlock(rem);
         if (m_rememberLine != 0)
         {
-            kDebug() << "%%%%%%%%%%%%%%%%% m_rememberLine still set!";
             // this probably means we had a block containing only 0x2029, so Scribe merged the userData/userState into the next
             m_rememberLine = 0;
         }
     }
 
     m_rememberLine = appendLine(IRCView::RememberLine);
-
-    kDebug() << (void*)m_rememberLine;
 }
 
 void IRCView::voidLineBlock(QTextBlock rem)
@@ -524,9 +449,6 @@ void IRCView::voidLineBlock(QTextBlock rem)
 
 void IRCView::clearLines()
 {
-    DebugBanner;
-    kDebug() << document();
-
     while (hasLines())
     {
         //IRCView::blockDeleted takes care of the pointers
@@ -541,17 +463,11 @@ void IRCView::wipeLineParagraphs()
 
 bool IRCView::hasLines()
 {
-    DebugBanner;
-
-    kDebug() << (void*)m_lastMarkerLine << (void*)m_rememberLine;
-
     return m_lastMarkerLine != 0;
 }
 
 Burr* IRCView::appendLine(IRCView::ObjectFormats type)
 {
-    DebugBanner;
-
     ScrollBarPin barpin(verticalScrollBar());
     SelectionPin selpin(this);
 
@@ -570,8 +486,6 @@ Burr* IRCView::appendLine(IRCView::ObjectFormats type)
 
     //TODO figure out what this is for
     cursor.setPosition(block.position());
-
-    kDebug() << block;
 
     return b;
 }
@@ -2038,7 +1952,7 @@ void IRCView::mouseReleaseEvent(QMouseEvent *ev)
         {
             // The QUrl magic is what QTextBrowser's anchorClicked() does internally;
             // we copy it here for consistent behavior between left and middle clicks.
-            openLink(QUrl::fromEncoded(m_urlToCopy.toUtf8()));
+            openLink(QUrl::fromEncoded(m_urlToCopy.toUtf8())); // krazy:exclude=qclasses
             return;
         }
         else
